fix: finishing refactoring
