+++ lib/cripto_details/view/detail_screen.dart
@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../shared/providers/wallet_providers.dart';
-import '../cripto_convert/convert_screen.dart';
-import '../shared/widgets/default_appbar.dart';
-import '../shared/widgets/default_subtitle.dart';
-import '../shared/widgets/default_title.dart';
-import '../shared/widgets/number_formatter.dart';
-import 'widgets/item_detail.dart';
+import '../../cripto_convert/view/convert_screen.dart';
+import '../../shared/providers/wallet_providers.dart';
+import '../../shared/widgets/default_appbar.dart';
+import '../../shared/widgets/default_subtitle.dart';
+import '../../shared/widgets/default_title.dart';
+import '../../shared/widgets/number_formatter.dart';
+import '../widgets/item_detail.dart';
 
 class DetailScreen extends StatefulHookConsumerWidget {
   const DetailScreen({Key? key}) : super(key: key);
