+++ lib/wallet/widgets/cripto_item.dart
@@ -1,13 +1,12 @@
 import 'package:decimal/decimal.dart';
 import 'package:decimal/intl.dart';
-
-import '../../shared/models/cripto_model.dart';
-import '../../shared/providers/wallet_providers.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../cripto_details/detail_screen.dart';
+import '../../cripto_details/view/detail_screen.dart';
+import '../../shared/models/cripto_model.dart';
+import '../../shared/providers/wallet_providers.dart';
 import '../../shared/widgets/number_formatter.dart';
 
 class CriptoItem extends StatefulHookConsumerWidget {
